# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * AI Investment API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface MarketAnalysisSource {
  title: string;
  url: string;
  published?: string;
}

export type MarketAnalysisMarketSentiment = typeof MarketAnalysisMarketSentiment[keyof typeof MarketAnalysisMarketSentiment];


export const MarketAnalysisMarketSentiment = {
  Positive: 'Positive',
  Neutral: 'Neutral',
  Negative: 'Negative',
} as const;

export type MarketAnalysisRiskLevel = typeof MarketAnalysisRiskLevel[keyof typeof MarketAnalysisRiskLevel];


export const MarketAnalysisRiskLevel = {
  Low: 'Low',
  Moderate: 'Moderate',
  High: 'High',
} as const;

export interface MarketAnalysis {
  summary: string;
  marketSentiment: MarketAnalysisMarketSentiment;
  riskLevel: MarketAnalysisRiskLevel;
  positiveFactors: string[];
  negativeFactors: string[];
  strongSectors: string[];
  weakSectors: string[];
  keyRisks: string[];
  sources: MarketAnalysisSource[];
  timestamp: string;
  /** Time taken to complete the analysis in milliseconds */
  analysisDuration: number;
}

/**
 * The structured analysis result produced by the module
 */
export type RepositoryEntryResult = { [key: string]: unknown };

/**
 * A module's latest stored analysis result with metadata
 */
export interface RepositoryEntry {
  /** Stable identifier for the module, e.g. "market-monitor" */
  moduleName: string;
  /** The structured analysis result produced by the module */
  result: RepositoryEntryResult;
  /** ISO 8601 — when this module first saved a result */
  createdAt: string;
  /** ISO 8601 — when this module last saved a result */
  updatedAt: string;
}

export interface ErrorResponse {
  error: string;
}