# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AI Investment API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: market-monitor
    description: Market Monitor analysis
  - name: repository
    description: Analysis Repository — shared read interface for all module results
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /market-monitor/analyze:
    post:
      operationId: runMarketAnalysis
      tags: [market-monitor]
      summary: Run market analysis
      description: Calls the AI service to analyze current market conditions and returns structured JSON
      responses:
        "200":
          description: Market analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarketAnalysis"
        "500":
          description: Analysis failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /repository:
    get:
      operationId: listRepositoryEntries
      tags: [repository]
      summary: List all stored module results
      description: Returns every module's latest analysis result, ordered by most recently updated
      responses:
        "200":
          description: All stored entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RepositoryEntry"
  /repository/{module}:
    get:
      operationId: getRepositoryEntry
      tags: [repository]
      summary: Get a specific module's latest result
      parameters:
        - name: module
          in: path
          required: true
          schema:
            type: string
          description: Module identifier, e.g. "market-monitor"
      responses:
        "200":
          description: The module's latest stored entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RepositoryEntry"
        "404":
          description: No result stored for this module
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    MarketAnalysisSource:
      type: object
      properties:
        title:
          type: string
        url:
          type: string
        published:
          type: string
      required:
        - title
        - url
    MarketAnalysis:
      type: object
      properties:
        summary:
          type: string
        marketSentiment:
          type: string
          enum: [Positive, Neutral, Negative]
        riskLevel:
          type: string
          enum: [Low, Moderate, High]
        positiveFactors:
          type: array
          items:
            type: string
        negativeFactors:
          type: array
          items:
            type: string
        strongSectors:
          type: array
          items:
            type: string
        weakSectors:
          type: array
          items:
            type: string
        keyRisks:
          type: array
          items:
            type: string
        sources:
          type: array
          items:
            $ref: "#/components/schemas/MarketAnalysisSource"
        timestamp:
          type: string
        analysisDuration:
          type: number
          description: Time taken to complete the analysis in milliseconds
      required:
        - summary
        - marketSentiment
        - riskLevel
        - positiveFactors
        - negativeFactors
        - strongSectors
        - weakSectors
        - keyRisks
        - sources
        - timestamp
        - analysisDuration
    RepositoryEntry:
      type: object
      description: A module's latest stored analysis result with metadata
      properties:
        moduleName:
          type: string
          description: Stable identifier for the module, e.g. "market-monitor"
        result:
          type: object
          description: The structured analysis result produced by the module
          additionalProperties: true
        createdAt:
          type: string
          description: ISO 8601 — when this module first saved a result
        updatedAt:
          type: string
          description: ISO 8601 — when this module last saved a result
      required:
        - moduleName
        - result
        - createdAt
        - updatedAt
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error