# lib/api-zod/src/generated/types/repositoryEntry.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * AI Investment API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RepositoryEntryResult } from './repositoryEntryResult';

/**
 * A module's latest stored analysis result with metadata
 */
export interface RepositoryEntry {
  /** Stable identifier for the module, e.g. "market-monitor" */
  moduleName: string;
  /** The structured analysis result produced by the module */
  result: RepositoryEntryResult;
  /** ISO 8601 — when this module first saved a result */
  createdAt: string;
  /** ISO 8601 — when this module last saved a result */
  updatedAt: string;
}